+++ Yuusuke-Ootsuji-educure-app/src/main/java/jp/co/example/service/RoleService.java
@@ -12,7 +12,7 @@ public class RoleService {
     @Autowired
     private RoleDao roleDao;
 
-    public Role getRoleById(Long roleId) {
+    public Role getRoleById(int roleId) {
         return roleDao.findById(roleId).orElse(null);
     }
 }
